moved rules to module
